requirements-mac.txt
pip
psutil==5.9.0
pillow
pyserial
colorama
simplejson
urllib3
selenium
python-dateutil
requests
wheel
pyautogui
Appium-Python-Client
uiautomator
lxml
tzlocal
xlwings
image
tzlocal
pyautocad
PyPDF2
pyshortcuts
datefinder
regex
pyttsx3
imutils
https://github.com/AutomationSolutionz/PyGetWindow-0.0.5/archive/master.zip
appscript
webdriver-manager
jsonschema
webcolors
jsonpointer
rfc3339-validator
rfc3986-validator
idna
pyrsistent
attrs
pytest==6.1.2
yapf
websocket-client==1.3.3
tqdm
pyscreenshot
pyyaml
imap-tools
keyboard
mouse
rich
beautifulsoup4
premailer
html-diff
pytz
rauth
python-dotenv
google-cloud-bigquery
google-cloud-bigquery-storage
deepdiff
xmltodict
configobj
jinja2
pandas
pyperclip
thefuzz
genson
google-cloud-storage
mitmproxy